backend/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
opencv-python==4.8.1.78
ultralytics==8.0.196
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Flask-CORS==4.0.0
prisma==0.12.0
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3 